requirements.txt
flask==2.3.3
flask-jwt-extended==4.4.4
gunicorn==20.1.0
pygit2==1.17.0
requests==2.31.0
pymongo==4.6.0
dnspython==2.4.2
python-dotenv==1.0.0
bcrypt==4.0.1
flask-cors==4.0.1